requirements.txt
mcp>=0.9.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0
azure-keyvault-certificates>=4.8.0
cryptography>=41.0.0